requirements.txt
streamlit
qdrant-client
mixedbread
python-dotenv